ML model trained on data collected from webcam.
This is the ML model created using Support Vector Machine (SVM) algorithm. It was really interesting working on this project as I got learn about sklearn, SVM, maximisation and optimisation, hyperplanes etc. However, the limitations of this program are: (1) Small, non-generalised datasets and (2) Slow performance. 
But this is a personal project, so I've decided to look past this.